refactor: merge useContentTaxonomyTagsCount and useContentTagsCount
