+++ src/content-tags-drawer/data/apiHooks.jsx
@@ -11,7 +11,6 @@ import {
   getContentTaxonomyTagsData,
   getContentData,
   updateContentTaxonomyTags,
-  getContentTaxonomyTagsCount,
 } from './api';
 
 /** @typedef {import("../../taxonomy/tag-list/data/types.mjs").TagListData} TagListData */
@@ -107,18 +106,6 @@ export const useContentTaxonomyTagsData = (contentId) => (
 );
 
 /**
- * Build the query to get the count og taxonomy tags applied to the content object
- * @param {string} contentId The ID of the content object to fetch the count of the applied tags for
- */
-// FixMe: remove
-// export const useContentTaxonomyTagsCount = (contentId) => (
-//   useQuery({
-//     queryKey: ['contentTaxonomyTagsCount', contentId],
-//     queryFn: () => getContentTaxonomyTagsCount(contentId),
-//   })
-// );
-
-/**
  * Builds the query to get meta data about the content object
  * @param {string} contentId The id of the content object (unit/component)
  */
@@ -159,7 +146,6 @@ export const useContentTaxonomyTagsUpdater = (contentId, taxonomyId) => {
       }
       queryClient.invalidateQueries({ queryKey: ['contentTagsCount', contentPattern] });
       // FixMe: remove code below
-      // queryClient.invalidateQueries({ queryKey: ['unitTagsCount'] });
       // queryClient.invalidateQueries({ queryKey: ['contentTaxonomyTagsCount', contentId] });
     },
   });
