+++ src/content-tags-drawer/tags-sidebar-controls/TagsSidebarHeader.jsx
@@ -4,8 +4,8 @@ import { Stack } from '@openedx/paragon';
 import { useParams } from 'react-router-dom';
 import { useIntl } from '@edx/frontend-platform/i18n';
 
+import { useContentTagsCount } from '../../generic/data/apiHooks';
 import messages from '../messages';
-import { useContentTaxonomyTagsCount } from '../data/apiHooks';
 import TagCount from '../../generic/tag-count';
 
 const TagsSidebarHeader = () => {
@@ -13,9 +13,9 @@ const TagsSidebarHeader = () => {
   const contentId = useParams().blockId;
 
   const {
-    data: contentTaxonomyTagsCount,
-    isSuccess: isContentTaxonomyTagsCountLoaded,
-  } = useContentTaxonomyTagsCount(contentId || '');
+    data: contentTagsCount,
+    isSuccess: isContentTagsCountLoaded,
+  } = useContentTagsCount(contentId || '');
 
   return (
     <Stack
@@ -25,8 +25,8 @@ const TagsSidebarHeader = () => {
       <h3 className="course-unit-sidebar-header-title m-0">
         {intl.formatMessage(messages.tagsSidebarTitle)}
       </h3>
-      { isContentTaxonomyTagsCountLoaded
-        && <TagCount count={contentTaxonomyTagsCount} />}
+      { isContentTagsCountLoaded
+        && <TagCount count={contentTagsCount} />}
     </Stack>
   );
 };
