+++ src/content-tags-drawer/tags-sidebar-controls/TagsSidebarHeader.test.jsx
@@ -1,36 +1,39 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { IntlProvider } from '@edx/frontend-platform/i18n';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import TagsSidebarHeader from './TagsSidebarHeader';
-import { useContentTaxonomyTagsCount } from '../data/apiHooks';
 
-jest.mock('../data/apiHooks', () => ({
-  useContentTaxonomyTagsCount: jest.fn(() => ({
-    isSuccess: false,
-    data: 17,
-  })),
+const mockGetTagsCount = jest.fn();
+
+jest.mock('../../generic/data/api', () => ({
+  ...jest.requireActual('../../generic/data/api'),
+  getTagsCount: () => mockGetTagsCount(),
 }));
+
+jest.mock('react-router-dom', () => ({
+  ...jest.requireActual('react-router-dom'),
+  useParams: () => ({ blockId: '123' }),
+}));
+
+const queryClient = new QueryClient();
 
 const RootWrapper = () => (
   <IntlProvider locale="en" messages={{}}>
-    <TagsSidebarHeader />
+    <QueryClientProvider client={queryClient}>
+      <TagsSidebarHeader />
+    </QueryClientProvider>
   </IntlProvider>
 );
 
 describe('<TagsSidebarHeader>', () => {
-  it('should not render count on loading', () => {
+  it('should render count only after query is complete', async () => {
+    let resolvePromise;
+    mockGetTagsCount.mockReturnValueOnce(new Promise((resolve) => { resolvePromise = resolve; }));
     render(<RootWrapper />);
     expect(screen.getByRole('heading', { name: /unit tags/i })).toBeInTheDocument();
     expect(screen.queryByText('17')).not.toBeInTheDocument();
-  });
-
-  it('should render count after query is complete', () => {
-    useContentTaxonomyTagsCount.mockReturnValue({
-      isSuccess: true,
-      data: 17,
-    });
-    render(<RootWrapper />);
-    expect(screen.getByRole('heading', { name: /unit tags/i })).toBeInTheDocument();
-    expect(screen.getByText('17')).toBeInTheDocument();
+    resolvePromise({ 123: 17 });
+    expect(await screen.findByText('17')).toBeInTheDocument();
   });
 });
