+++ src/course-outline/__mocks__/index.js
@@ -4,4 +4,3 @@ export { default as courseBestPracticesMock } from './courseBestPractices';
 export { default as courseLaunchMock } from './courseLaunch';
 export { default as courseSectionMock } from './courseSection';
 export { default as courseSubsectionMock } from './courseSubsection';
-export { default as contentTagsCountMock } from './contentTagsCount';
