+++ src/generic/__mocks__/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as contentTagsCountMock } from './contentTagsCount';
