+++ src/generic/data/api.test.js
@@ -2,6 +2,7 @@ import MockAdapter from 'axios-mock-adapter';
 import { initializeMockApp } from '@edx/frontend-platform';
 import { getAuthenticatedHttpClient } from '@edx/frontend-platform/auth';
 
+import { contentTagsCountMock } from '../__mocks__';
 import {
   createOrRerunCourse,
   getApiBaseUrl,
@@ -9,6 +10,8 @@ import {
   getCreateOrRerunCourseUrl,
   getCourseRerunUrl,
   getCourseRerun,
+  getTagsCount,
+  getTagsCountApiUrl,
 } from './api';
 
 let axiosMock;
@@ -71,5 +74,20 @@ describe('generic api calls', () => {
 
     expect(axiosMock.history.post[0].url).toEqual(getCreateOrRerunCourseUrl());
     expect(result).toEqual(courseRerunData);
+  });
+
+  it('should get tags count', async () => {
+    const pattern = 'this,is,a,pattern';
+    const contentId = 'block-v1:SampleTaxonomyOrg1+STC1+2023_1+type@vertical+block@aaf8b8eb86b54281aeeab12499d2cb06';
+    axiosMock.onGet().reply(200, contentTagsCountMock);
+    const result = await getTagsCount(pattern);
+    expect(axiosMock.history.get[0].url).toEqual(getTagsCountApiUrl(pattern));
+    expect(result).toEqual(contentTagsCountMock);
+    expect(contentTagsCountMock[contentId]).toEqual(15);
+  });
+
+  it('should get null on empty pattern', async () => {
+    const result = await getTagsCount('');
+    expect(result).toEqual(null);
   });
 });
